# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Finchat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.finchat.com",
    "basePath": "/",
    "paths": {
        "/auth/v1/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login a user",
                "operationId": "login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httperr.HTTPErr"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httperr.HTTPErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httperr.HTTPErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httperr.HTTPErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httperr.HTTPErr": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Application specific code.",
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "description": "User readable message.",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status.",
                    "type": "integer"
                }
            }
        }
    }
}